将Winform、WPF共用的抽象到YourWeather.Rcl.Desktop, 将Blazor的Wasm和SSR共用的抽象到YourWeather.Rcl.Web
