+++ YourWeather.sln
@@ -22,7 +22,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YourWeather.Shared", "YourW
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YourWeather.Winform", "YourWeather\YourWeather.Winform\YourWeather.Winform.csproj", "{5024BB5A-F9D3-4685-BD66-A5BA79EE2802}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YourWeather.Wpf", "YourWeather\YourWeather.Wpf\YourWeather.Wpf.csproj", "{1DE9AD1B-5976-4984-9B7C-0C066D0495C8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YourWeather.Wpf", "YourWeather\YourWeather.Wpf\YourWeather.Wpf.csproj", "{1DE9AD1B-5976-4984-9B7C-0C066D0495C8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YourWeather.Rcl.Web", "YourWeather.Rcl.Web\YourWeather.Rcl.Web.csproj", "{B9BA9FB7-70B4-4559-8086-3270AC783A6A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YourWeather.Rcl.Desktop", "YourWeather.Rcl.Desktop\YourWeather.Rcl.Desktop.csproj", "{46885133-7858-42E5-8C7E-83171520CCE9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -60,6 +64,14 @@ Global
 		{1DE9AD1B-5976-4984-9B7C-0C066D0495C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1DE9AD1B-5976-4984-9B7C-0C066D0495C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1DE9AD1B-5976-4984-9B7C-0C066D0495C8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B9BA9FB7-70B4-4559-8086-3270AC783A6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B9BA9FB7-70B4-4559-8086-3270AC783A6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B9BA9FB7-70B4-4559-8086-3270AC783A6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B9BA9FB7-70B4-4559-8086-3270AC783A6A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{46885133-7858-42E5-8C7E-83171520CCE9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{46885133-7858-42E5-8C7E-83171520CCE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{46885133-7858-42E5-8C7E-83171520CCE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{46885133-7858-42E5-8C7E-83171520CCE9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
