+++ YourWeather/README.md
@@ -1,0 +1,11 @@
+# 项目结构
+
+- YourWeather `Maui Blazor`
+- YourWeather.Client `Blazor WebAssembly`
+- YourWeather.Rcl `Razor类库，存放页面和一些基础实现`
+- YourWeather.Rcl.Desktop `Winform和WPF共用的一些实现类`
+- YourWeather.Rcl.Web `WebAssembly 和 Server 共用的一些实现类`
+- YourWeather.Server `Blazor Server`
+- YourWeather.Shared `共用的接口和模型`
+- YourWeather.Winform `Winform Blazor`
+- YourWeather.Wpf `WPF Blazor`
