+++ YourWeather/YourWeather.Client/YourWeather.Client.csproj
@@ -14,15 +14,16 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\YourWeather.Rcl\YourWeather.Rcl.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ServiceWorker Include="wwwroot\service-worker.js" PublishedContent="wwwroot\service-worker.published.js" />
   </ItemGroup>
 
   <ItemGroup>
     <Folder Include="Extend\ServiceCollectionExtend\" />
+    <Folder Include="Services\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\YourWeather.Rcl.Web\YourWeather.Rcl.Web.csproj" />
   </ItemGroup>
 
 </Project>
