+++ YourWeather/YourWeather.Rcl.Desktop/Services/LocationService.cs
@@ -1,11 +1,12 @@
 ﻿using Darnton.Blazor.DeviceInterop.Geolocation;
 
-namespace YourWeather.Client.Services
+namespace YourWeather.Rcl.Desktop.Services
 {
     public class LocationService : Rcl.Services.LocationService
     {
         public LocationService(IGeolocationService geolocationService) : base(geolocationService)
         {
+
         }
     }
 }
