+++ YourWeather/YourWeather.Rcl.Desktop/Services/WeatherService.cs
@@ -1,4 +1,4 @@
-﻿namespace YourWeather.Server.Services
+﻿namespace YourWeather.Rcl.Desktop.Services
 {
     public class WeatherService : Rcl.Services.WeatherService
     {
