+++ YourWeather/YourWeather.Rcl.Web/Services/LocationService.cs
@@ -1,11 +1,12 @@
 ﻿using Darnton.Blazor.DeviceInterop.Geolocation;
 
-namespace YourWeather.Server.Services
+namespace YourWeather.Rcl.Web.Services
 {
     public class LocationService : Rcl.Services.LocationService
     {
         public LocationService(IGeolocationService geolocationService) : base(geolocationService)
         {
+
         }
     }
 }
