+++ YourWeather/YourWeather.Rcl.Web/Services/ThemeService.cs
@@ -1,4 +1,4 @@
-﻿namespace YourWeather.Client.Services
+﻿namespace YourWeather.Rcl.Web.Services
 {
     public class ThemeService : Rcl.Services.ThemeService
     {
