+++ YourWeather/YourWeather.Rcl.Web/_Imports.razor
@@ -1,0 +1,1 @@
+﻿@using Microsoft.AspNetCore.Components.Web
