--- YourWeather/YourWeather.Server/Services/SystemService.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.JSInterop;
-
-namespace YourWeather.Server.Services
-{
-    public class SystemService : Rcl.Services.SystemService
-    {
-        public SystemService(IJSRuntime js) : base(js)
-        {
-        }
-    }
-}
