--- YourWeather/YourWeather.Server/Services/ThemeService.cs
@@ -1,6 +1,0 @@
-﻿namespace YourWeather.Server.Services
-{
-    public class ThemeService : Rcl.Services.ThemeService
-    {
-    }
-}
