+++ YourWeather/YourWeather.Winform/Extend/ServiceCollectionExtend/AddCustomIOC.cs
@@ -1,6 +1,6 @@
-﻿using YourWeather.Winform.Services;
+﻿using Microsoft.Extensions.DependencyInjection;
+using YourWeather.Rcl.Desktop.Services;
 using YourWeather.Shared;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace YourWeather.Winform.Extend
 {
