--- YourWeather/YourWeather.Winform/Services/ThemeService.cs
@@ -1,11 +1,0 @@
-﻿namespace YourWeather.Winform.Services
-{
-    public class ThemeService : Rcl.Services.ThemeService
-    {
-        protected override void NotifyStateChanged()
-        {
-            base.NotifyStateChanged();
-            TitleBar.EnableDarkMode(Dark);
-        }
-    }
-}
