--- YourWeather/YourWeather.Winform/Services/WeatherService.cs
@@ -1,6 +1,0 @@
-﻿namespace YourWeather.Winform.Services
-{
-    public class WeatherService : Rcl.Services.WeatherService
-    {
-    }
-}
