--- YourWeather/YourWeather.Wpf/Services/WeatherService.cs
@@ -1,6 +1,0 @@
-﻿namespace YourWeather.Wpf.Services
-{
-    public class WeatherService : Rcl.Services.WeatherService
-    {
-    }
-}
