+++ YourWeather/YourWeather.Wpf/YourWeather.Wpf.csproj
@@ -24,17 +24,18 @@
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore.Components.WebView.Wpf" Version="7.0.81" />
-		<PackageReference Include="Blazored.LocalStorage" Version="4.3.0" />
-		<PackageReference Include="PInvoke.User32" Version="0.7.124" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\YourWeather.Rcl\YourWeather.Rcl.csproj" />
-	  <ProjectReference Include="..\YourWeather.Shared\YourWeather.Shared.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
 	  <Resource Include="appiconfg.svg" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <Folder Include="Services\" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\..\YourWeather.Rcl.Desktop\YourWeather.Rcl.Desktop.csproj" />
 	</ItemGroup>
 
 </Project>
